Remove Renderers dependency on InteractionHandler
This separates them even further, so they should work independent of each other in theory
