+++ spigot/src/main/kotlin/com/wolfyscript/viewportl/spigot/gui/inventoryui/interaction/InventoryUIListener.kt
@@ -17,6 +17,9 @@
  */
 package com.wolfyscript.viewportl.spigot.gui.inventoryui.interaction
 
+import com.wolfyscript.scafall.spigot.api.wrappers.wrap
+import com.wolfyscript.viewportl.common.gui.reactivity.ReactiveGraph
+import com.wolfyscript.viewportl.gui.ViewRuntime
 import com.wolfyscript.viewportl.spigot.gui.inventoryui.BukkitInventoryGuiHolder
 import org.bukkit.event.EventHandler
 import org.bukkit.event.EventPriority
@@ -25,14 +28,20 @@ import org.bukkit.event.inventory.InventoryClickEvent
 import org.bukkit.event.inventory.InventoryCloseEvent
 import org.bukkit.event.inventory.InventoryDragEvent
 
-class InventoryUIListener : Listener {
+class InventoryUIListener(val runtime: ViewRuntime<*, SpigotInvUIInteractionHandler>) : Listener {
 
     @EventHandler(priority = EventPriority.HIGHEST)
     fun onInvClick(event: InventoryClickEvent) {
         val holder = event.inventory.holder
         if (holder is BukkitInventoryGuiHolder) {
 //            bukkitInventoryGuiHolder.guiHolder().getViewManager().blockedByInteraction();
-            holder.onClick(event)
+            val details = ClickInteractionDetailsImpl(event)
+            runtime.interactionHandler.onClick(details)
+            event.isCancelled = true
+
+            runtime.viewportl.scafall.scheduler.syncTask(runtime.viewportl.scafall.corePlugin) {
+                (runtime.reactiveSource as ReactiveGraph).runEffects()
+            }
         }
     }
 
@@ -41,7 +50,30 @@ class InventoryUIListener : Listener {
         val holder = event.inventory.holder
         if (holder is BukkitInventoryGuiHolder) {
 //            bukkitInventoryGuiHolder.guiHolder().getViewManager().blockedByInteraction();
-            holder.onDrag(event)
+            if (event.rawSlots.stream()
+                    .anyMatch { rawSlot -> event.view.getInventory(rawSlot) != event.view.topInventory }
+            ) {
+                event.isCancelled = true
+                return
+            }
+            if (runtime.window == null) return
+            val interactionDetails = DragInteractionDetailsImpl(event)
+            runtime.interactionHandler.onDrag(interactionDetails)
+            if (!interactionDetails.valid) {
+                event.isCancelled = true
+            } else {
+                runtime.viewportl.scafall.scheduler.syncTask(runtime.viewportl.scafall.corePlugin) {
+                    for (rawSlot in event.rawSlots) {
+                        if (rawSlot < event.inventory.size) {
+                            interactionDetails.callSlotValueUpdate(rawSlot, event.inventory.getItem(rawSlot)?.wrap())
+                        }
+                    }
+                }
+            }
+
+            runtime.viewportl.scafall.scheduler.syncTask(runtime.viewportl.scafall.corePlugin) {
+                (runtime.reactiveSource as ReactiveGraph).runEffects()
+            }
         }
     }
 
@@ -50,7 +82,7 @@ class InventoryUIListener : Listener {
         val holder = event.inventory.holder
         if (holder is BukkitInventoryGuiHolder) {
 //            bukkitInventoryGuiHolder.guiHolder().getViewManager().blockedByInteraction();
-            holder.onClose(event)
+            runtime.window?.apply { close() }
         }
     }
 }
